+++ src/core/app.module.core.ts
@@ -2,12 +2,14 @@ import { ConsoleLogger, Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { PassportModule } from '@nestjs/passport';
 import { ServeStaticModule } from '@nestjs/serve-static';
+import { TerminusModule } from '@nestjs/terminus';
 
 import { AuthController } from '../api/auth.controller';
 import { ChatsController } from '../api/chats.controller';
 import { ChattingController } from '../api/chatting.controller';
 import { ContactsController } from '../api/contacts.controller';
 import { GroupsController } from '../api/groups.controller';
+import { HealthController } from '../api/health.controller';
 import { PresenceController } from '../api/presence.controller';
 import { ScreenshotController } from '../api/screenshot.controller';
 import {
@@ -18,6 +20,8 @@ import { StatusController } from '../api/status.controller';
 import { VersionController } from '../api/version.controller';
 import { WhatsappConfigService } from '../config.service';
 import { SessionManager } from './abc/manager.abc';
+import { WAHAHealthCheckService } from './abc/WAHAHealthCheckService';
+import { WAHAHealthCheckServiceCore } from './health/WAHAHealthCheckServiceCore';
 import { SessionManagerCore } from './manager.core';
 
 export const IMPORTS = [
@@ -38,6 +42,7 @@ export const IMPORTS = [
     },
   }),
   PassportModule,
+  TerminusModule,
 ];
 export const CONTROLLERS = [
   AuthController,
@@ -51,11 +56,16 @@ export const CONTROLLERS = [
   PresenceController,
   ScreenshotController,
   VersionController,
+  HealthController,
 ];
 const PROVIDERS = [
   {
     provide: SessionManager,
     useClass: SessionManagerCore,
+  },
+  {
+    provide: WAHAHealthCheckService,
+    useClass: WAHAHealthCheckServiceCore,
   },
   WhatsappConfigService,
   ConsoleLogger,